requirements/base.txt
annotated-types==0.7.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
backports.asyncio.runner==1.2.0
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
coverage==7.13.5
cryptography==47.0.0
daphne==4.2.1
dj-database-url==3.1.2
Django==5.2.13
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
execnet==2.1.2
factory_boy==3.3.3
Faker==40.13.0
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
psycopg==3.3.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.1.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-django==4.12.0
pytest-randomly==4.0.1
pytest-xdist==3.8.0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.4.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
zope.interface==8.4
